Moved Firestore internals to different package
