+++ firebase-firestore/src/androidMain/kotlin/dev/gitlive/firebase/firestore/internal/SetOptions.kt
@@ -1,0 +1,8 @@
+package dev.gitlive.firebase.firestore.internal
+
+internal val SetOptions.android: com.google.firebase.firestore.SetOptions? get() = when (this) {
+    is SetOptions.Merge -> com.google.firebase.firestore.SetOptions.merge()
+    is SetOptions.Overwrite -> null
+    is SetOptions.MergeFields -> com.google.firebase.firestore.SetOptions.mergeFields(fields)
+    is SetOptions.MergeFieldPaths -> com.google.firebase.firestore.SetOptions.mergeFieldPaths(encodedFieldPaths)
+}
