+++ firebase-firestore/src/commonMain/kotlin/dev/gitlive/firebase/firestore/DocumentReferenceSerializer.kt
@@ -1,5 +1,6 @@
 package dev.gitlive.firebase.firestore
 
+import dev.gitlive.firebase.firestore.internal.NativeDocumentReference
 import dev.gitlive.firebase.internal.FirebaseEncoder
 import dev.gitlive.firebase.internal.SpecialValueSerializer
 import kotlinx.serialization.KSerializer
