+++ firebase-firestore/src/commonMain/kotlin/dev/gitlive/firebase/firestore/internal/NativeFirebaseFirestoreWrapper.kt
@@ -1,0 +1,24 @@
+package dev.gitlive.firebase.firestore.internal
+
+import dev.gitlive.firebase.firestore.FirebaseFirestoreSettings
+import dev.gitlive.firebase.firestore.NativeCollectionReference
+import dev.gitlive.firebase.firestore.NativeFirebaseFirestore
+import dev.gitlive.firebase.firestore.NativeQuery
+import dev.gitlive.firebase.firestore.NativeTransaction
+import dev.gitlive.firebase.firestore.NativeWriteBatch
+
+internal expect class NativeFirebaseFirestoreWrapper internal constructor(native: NativeFirebaseFirestore) {
+    val native: NativeFirebaseFirestore
+    var settings: FirebaseFirestoreSettings
+
+    fun collection(collectionPath: String): NativeCollectionReference
+    fun collectionGroup(collectionId: String): NativeQuery
+    fun document(documentPath: String): NativeDocumentReference
+    fun batch(): NativeWriteBatch
+    fun setLoggingEnabled(loggingEnabled: Boolean)
+    suspend fun clearPersistence()
+    suspend fun <T> runTransaction(func: suspend NativeTransaction.() -> T): T
+    fun useEmulator(host: String, port: Int)
+    suspend fun disableNetwork()
+    suspend fun enableNetwork()
+}
