+++ firebase-firestore/src/commonMain/kotlin/dev/gitlive/firebase/firestore/internal/SafeValue.kt
@@ -1,0 +1,14 @@
+package dev.gitlive.firebase.firestore.internal
+
+import dev.gitlive.firebase.firestore.DocumentReference
+import dev.gitlive.firebase.firestore.GeoPoint
+import dev.gitlive.firebase.firestore.Timestamp
+
+internal val Any.safeValue: Any get() = when (this) {
+    is Timestamp -> nativeValue
+    is GeoPoint -> nativeValue
+    is DocumentReference -> native.nativeValue
+    is Map<*, *> -> this.mapNotNull { (key, value) -> key?.let { it.safeValue to value?.safeValue } }
+    is Collection<*> -> this.mapNotNull { it?.safeValue }
+    else -> this
+}
