+++ firebase-firestore/src/commonMain/kotlin/dev/gitlive/firebase/firestore/internal/SetOptions.kt
@@ -1,0 +1,13 @@
+package dev.gitlive.firebase.firestore.internal
+
+import dev.gitlive.firebase.firestore.FieldPath
+
+@PublishedApi
+internal sealed class SetOptions {
+    data object Merge : SetOptions()
+    data object Overwrite : SetOptions()
+    data class MergeFields(val fields: List<String>) : SetOptions()
+    data class MergeFieldPaths(val fieldPaths: List<FieldPath>) : SetOptions() {
+        val encodedFieldPaths = fieldPaths.map { it.encoded }
+    }
+}
