+++ firebase-firestore/src/jsMain/kotlin/dev/gitlive/firebase/firestore/internal/JSQueryGet.kt
@@ -1,0 +1,7 @@
+package dev.gitlive.firebase.firestore.internal
+
+import dev.gitlive.firebase.firestore.Source
+import dev.gitlive.firebase.firestore.externals.Query
+import dev.gitlive.firebase.firestore.externals.getDocs
+import dev.gitlive.firebase.firestore.externals.getDocsFromCache
+import dev.gitlive.firebase.firestore.externals.getDocsFromServer
